+++ src/config/components-config.ts
@@ -85,7 +85,16 @@ export interface PharosPanel {
   api?: PharosApi[];
 }
 
-const ppi_interaction_fields: PharosApi[] = [
+const commonGwasFields: PharosApi[] =  [
+  { field: 'studyCount', label: 'Study Count', description: 'Number of studies that have found the association between the gene and the trait'},
+  { field: 'snpCount', label: 'SNP Count', description: 'Number of SNPs involved in the association'},
+  { field: 'betaCount', label: 'Beta Count', description: 'Count of beta values with 95% confidence intervals that support the association'},
+  { field: 'medianOddsRatio', label: 'Odds Ratio', description: 'Median Odds Ratio that supports the association'},
+  { field: 'evidence', label: 'Evidence', description: 'Gene-trait pairs (GTs) are ranked based on selected variables, determined by benchmarking versus gold standard associations'},
+  { field: 'provenance', label: 'Provenance', description: 'Link to the TIGA page for study details and publications supporting the association'}
+];
+
+const ppiInteractionFields: PharosApi[] = [
   {
     field: 'Score',
     label: 'score',
@@ -108,7 +117,7 @@ const ppi_interaction_fields: PharosApi[] = [
   }
 ];
 
-const disease_association_fields: PharosApi[] =
+const diseaseAssociationFields: PharosApi[] =
   [
     {
       field: 'evidence',
@@ -157,7 +166,7 @@ const disease_association_fields: PharosApi[] =
     }
   ];
 
-const similarity_fields: PharosApi[] =
+const similarityFields: PharosApi[] =
   [
     {
       field: 'jaccard',
@@ -236,9 +245,9 @@ const TARGET_TABLE_COMPONENT: PharosPanel = {
       label: 'Novelty',
       description: 'Tin-X metric for the relative scarcity of specific publications for this target.'
     },
-    ...ppi_interaction_fields,
-    ...disease_association_fields,
-    ...similarity_fields,
+    ...ppiInteractionFields,
+    ...diseaseAssociationFields,
+    ...similarityFields,
     {
       field: 'illuminationGraph',
       label: 'Illumination Graph',
@@ -525,7 +534,7 @@ const DISEASE_SOURCE_PANEL: PharosPanel = {
       label: 'Disease ID',
       description: 'Identifier the data source is using for this disease association.'
     },
-    ...disease_association_fields,
+    ...diseaseAssociationFields,
     {
       field: 'dataSources',
       label: 'Data Source Links',
@@ -711,6 +720,25 @@ const ORTHOLOG_VARIANT_PANEL: PharosPanel = {
   ]
 };
 
+const GWAS_TARGET_ANALYTICS_PANEL: PharosPanel = {
+  token: TOKENS.GWAS_TARGET_ANALYTICS_PANEL,
+  navHeader: {
+    label: 'GWAS Traits',
+    section: 'tiga',
+    mainDescription: 'Genome-wide association studies (GWAS) find associations between phenotypic traits and genes. Target Illumination ' +
+      'GWAS Analytics (TIGA) scores and ranks those traits according to a subset of study parameters.',
+    mainSource: 'https://unmtid-shinyapps.net/shiny/tiga/'
+  }, api: [
+    { field: 'reference', label: 'Citation', description: 'TIGA scores and ranks traits (or genes) by meanRank based on three variables: ' +
+        '(1) pVal_mlog which is max(-Log(pValue)), (2) N_snpw, SNP count weighted by genomic distance, and (3) RCRAS, Relative Citation ' +
+        'Ratio Aggregated Score, based on iCite RCR. These variables were selected by benchmark against the gold standard target-phenotype associations.',
+      source: 'https://www.biorxiv.org/content/10.1101/2020.11.11.378596v2'},
+    { field: 'gwasTrait', label: 'GWAS Trait', description: 'The phenotypic trait found to be associated with the gene'},
+    { field: 'efoID', label: 'EFO ID', description: 'Experimental Factor Ontology (EFO) ID for the trait'},
+    ...commonGwasFields
+  ]
+};
+
 const GO_TERMS_PANEL: PharosPanel = {
   token: TOKENS.GO_TERMS_PANEL,
   navHeader: {
@@ -817,7 +845,7 @@ const PROTEIN_PROTEIN_PANEL: PharosPanel = {
       label: 'Novelty',
       description: 'Tin-X metric for the relative scarcity of specific publications for this target.'
     },
-    ...ppi_interaction_fields,
+    ...ppiInteractionFields,
     {
       field: 'Data Source',
       label: 'Data Sources',
@@ -987,7 +1015,7 @@ const TARGET_FACET_PANEL: PharosPanel = {
     },
     {
       field: 'mgiPhenotype',
-      label: "MGI Phenotype",
+      label: 'MGI Phenotype',
       description: 'Occurrence of this target in up to 10 categories of JAX/MGI Phenotypes.'
     }
   ]
@@ -1016,14 +1044,23 @@ const DISEASE_TABLE_COMPONENT: PharosPanel = {
 };
 
 /**
- * disease summary header component
+ * disease header component
  * @type {PharosPanel}
  */
 const DISEASE_HEADER_COMPONENT: PharosPanel = {
   token: TOKENS.DISEASE_HEADER_COMPONENT,
+  section: Position.Header
+};
+
+/**
+ * disease summary header component
+ * @type {PharosPanel}
+ */
+const DISEASE_SUMMARY_COMPONENT: PharosPanel = {
+  token: TOKENS.DISEASE_SUMMARY_COMPONENT,
   navHeader: {
     label: 'Disease Summary',
-    section: 'diseaseHeader',
+    section: 'diseaseSummary',
     mainDescription: 'High level summary of knowledge for a disease, including descriptions and datasource references.'
   },
   api: [
@@ -1075,6 +1112,26 @@ const DISEASE_DO_BROWSER_COMPONENT: PharosPanel = {
         'Total count of associated targets, and a breakdown of targets by Target Development Level is shown.'
     }
   ]
+};
+
+const DISEASE_GWAS_ANALYTICS_COMPONENT: PharosPanel = {
+  token: TOKENS.DISEASE_GWAS_ANALYTICS_COMPONENT,
+  navHeader: {
+  label: 'GWAS Targets',
+    section: 'tiga',
+    mainDescription: 'Genome-wide association studies (GWAS) find associations between phenotypic traits and genes. Target Illumination ' +
+      'GWAS Analytics (TIGA) scores and ranks those traits according to a subset of study parameters.',
+    mainSource: 'https://unmtid-shinyapps.net/shiny/tiga/'
+}, api: [
+    { field: 'reference', label: 'Citation', description: 'TIGA scores and ranks traits (or genes) by meanRank based on three variables: ' +
+        '(1) pVal_mlog which is max(-Log(pValue)), (2) N_snpw, SNP count weighted by genomic distance, and (3) RCRAS, Relative Citation ' +
+        'Ratio Aggregated Score, based on iCite RCR. These variables were selected by benchmark against the gold standard target-phenotype associations.',
+      source: 'https://www.biorxiv.org/content/10.1101/2020.11.11.378596v2'},
+    { field: 'gwasTrait', label: 'trait', description: 'The phenotypic trait found to be associated with the genes in the list.'},
+    { field: 'target', label: 'Target', description: 'The target found to be associated with the disease.'},
+    { field: 'tdl', label: 'TDL', description: 'The target development level for the associated target'},
+    ...commonGwasFields
+]
 };
 
 const DISEASE_TINX_COMPONENT: PharosPanel = {
@@ -1277,6 +1334,7 @@ export const COMPONENTSCONFIG: Map<string, any> = new Map<string, any>(
           DRUGS_PANEL,
           LIGANDS_PANEL,
           DISEASE_SOURCE_PANEL,
+          GWAS_TARGET_ANALYTICS_PANEL,
           PDB_PANEL,
           PATHWAYS_PANEL,
           GO_TERMS_PANEL,
@@ -1302,8 +1360,11 @@ export const COMPONENTSCONFIG: Map<string, any> = new Map<string, any>(
       },
       details: {
         components: [
+          PHAROS_SUBNAV_COMPONENT,
           PHAROS_HELPPANEL_COMPONENT,
           DISEASE_HEADER_COMPONENT,
+          DISEASE_SUMMARY_COMPONENT,
+          DISEASE_GWAS_ANALYTICS_COMPONENT,
           DISEASE_DO_BROWSER_COMPONENT,
           DISEASE_TINX_COMPONENT
         ]